gitrest: persist pre-computed full summaries in storage (#9523)
