+++ server/gitrest/src/app.ts
@@ -12,7 +12,7 @@ import split from "split";
 import winston from "winston";
 import { bindCorrelationId } from "@fluidframework/server-services-utils";
 import * as routes from "./routes";
-import { IRepositoryManagerFactory } from "./utils";
+import { IFileSystemManager, IRepositoryManagerFactory } from "./utils";
 
 /**
  * Basic stream logging interface for libraries that require a stream to pipe output to
@@ -23,6 +23,7 @@ const stream = split().on("data", (message) => {
 
 export function create(
     store: nconf.Provider,
+    fileSystemManager: IFileSystemManager,
     repositoryManagerFactory: IRepositoryManagerFactory,
 ) {
     // Express app configuration
@@ -55,7 +56,7 @@ export function create(
 
     app.use(cors());
 
-    const apiRoutes = routes.create(store, repositoryManagerFactory);
+    const apiRoutes = routes.create(store, fileSystemManager, repositoryManagerFactory);
     app.use(apiRoutes.git.blobs);
     app.use(apiRoutes.git.refs);
     app.use(apiRoutes.git.repos);
