+++ server/gitrest/src/runnerFactory.ts
@@ -10,7 +10,7 @@ import * as core from "@fluidframework/server-services-core";
 import { normalizePort } from "@fluidframework/server-services-utils";
 import { ExternalStorageManager } from "./externalStorageManager";
 import { GitrestRunner } from "./runner";
-import { IRepositoryManagerFactory, NodegitRepositoryManagerFactory } from "./utils";
+import { IFileSystemManager, IRepositoryManagerFactory, NodegitRepositoryManagerFactory } from "./utils";
 
 export class GitrestResources implements core.IResources {
     public webServerFactory: core.IWebServerFactory;
@@ -18,6 +18,7 @@ export class GitrestResources implements core.IResources {
     constructor(
         public readonly config: Provider,
         public readonly port: string | number,
+        public readonly fileSystemManager: IFileSystemManager,
         public readonly repositoryManagerFactory: IRepositoryManagerFactory) {
         this.webServerFactory = new services.BasicWebServerFactory();
     }
@@ -32,13 +33,21 @@ export class GitrestResourcesFactory implements core.IResourcesFactory<GitrestRe
         const port = normalizePort(process.env.PORT || "3000");
         const fileSystemManager = fsPromises;
         const externalStorageManager = new ExternalStorageManager(config);
-        const repositoryManagerFactory = new NodegitRepositoryManagerFactory(
-            config.get("storageDir"),
-            fileSystemManager,
-            externalStorageManager,
-        );
+        const storageDirectory = config.get("storageDir");
+        const gitLibrary: string | undefined = config.get("git:lib:name");
+        const getRepositoryManagerFactory = () => {
+            if (!gitLibrary || gitLibrary === "nodegit") {
+                return new NodegitRepositoryManagerFactory(
+                    storageDirectory,
+                    fileSystemManager,
+                    externalStorageManager,
+                );
+            }
+            throw new Error("Invalid git library name.");
+        };
+        const repositoryManagerFactory = getRepositoryManagerFactory();
 
-        return new GitrestResources(config, port, repositoryManagerFactory);
+        return new GitrestResources(config, port, fileSystemManager, repositoryManagerFactory);
     }
 }
 
@@ -48,6 +57,7 @@ export class GitrestRunnerFactory implements core.IRunnerFactory<GitrestResource
             resources.webServerFactory,
             resources.config,
             resources.port,
+            resources.fileSystemManager,
             resources.repositoryManagerFactory);
     }
 }
