+++ server/gitrest/src/test/routes.spec.ts
@@ -160,7 +160,7 @@ describe("GitRest", () => {
         let supertest: request.SuperTest<request.Test>;
         beforeEach(() => {
             const repoManagerFactory = getRepoManagerFactory();
-            const testApp = app.create(testUtils.defaultProvider, repoManagerFactory);
+            const testApp = app.create(testUtils.defaultProvider, fsPromises, repoManagerFactory);
             supertest = request(testApp);
         });
